examples/fxn-ag2/requirements.txt
# requirements.txt
pyautogen[vision]
solana
base58
requests
pandas
pillow
python-dotenv
aiohttp
flaml[automl]
